# improve logits bias