Summary:
Signed-off-by: Yu Guo <yuguo@meta.com>

use `torch.index_put_(., accumulate=True)` to update the logits tensor once.

Rollback Plan:

Differential Revision: D75833790