@njhill 2025-06-03
Thanks @yuguo68. FYI we are planning to replace most of the sampling param implementations with vectorized versions soon, see https://github.com/vllm-project/vllm/pull/13360/files#diff-d01f143e1af472f24af24842cb879907ce624e6e5c977935e944545240723529R122