@yuguo68 2025-06-03
> Add some test plan about how to verify the correctness?

I think the unittest https://github.com/vllm-project/vllm/blob/bdce64f2365b39335141f8efcb3a0a8ecc559153/tests/v1/sample/test_sampler.py#L446 should cover it. let me double check.