@yuguo68 2025-06-03
> @yuguo68 , if we have obvious gain, we can still land the PR first, since landing the official new design still need some more time.

sounds good, I will do a microbenchmark.