@yuguo68 2025-06-05
microbenchmark with script on H100

```

from typing import Optional, Union

import torch

def async_tensor_h2d(
    data: list,
    dtype: torch.dtype,
    target_device: Union[str, torch.device],
    pin_memory: bool,
) -> torch.Tensor:
    """Asynchronously create a tensor and copy it from host to device."""
    t = torch.tensor(data, dtype=dtype, pin_memory=pin_memory, device="cpu")
    return t.to(device=target_device, non_blocking=True)


def apply_logits_bias_opt(
    logits: torch.Tensor,
    logit_bias_list: list[Optional[dict[int, float]]],
) -> torch.Tensor:

    rows: list[int] = []
    cols: list[int] = []
    vals: list[float] = []

    # Get vocabulary size from logits
    vocab_size = logits.shape[-1]

    for i, logit_bias in enumerate(logit_bias_list):
        if logit_bias:
            for token_id, bias in logit_bias.items():
                # Check token_id bounds to ensure within vocabulary
                if token_id < 0 or token_id >= vocab_size:
                    raise ValueError(
                        f"token_id {token_id} in logit_bias contains "
                        f"out-of-vocab token id. Vocabulary size: "
                        f"{vocab_size}"
                    )
                rows.append(i)
                cols.append(token_id)
                vals.append(bias)

    if rows:
        indices = async_tensor_h2d([rows, cols], torch.int64,
                                    logits.device, True)
        values = async_tensor_h2d(vals, torch.float, logits.device,
                                    True)
        logits.index_put_(tuple(indices), values=values, accumulate=True)

    torch.cuda.synchronize()
    return logits


def apply_logits_bias(
    logits: torch.Tensor,
    logit_bias_list: list[Optional[dict[int, float]]],
) -> torch.Tensor:

    # Get vocabulary size from logits
    vocab_size = logits.shape[-1]

    for i, logit_bias in enumerate(logit_bias_list):
        if logit_bias:
            for token_id, bias in logit_bias.items():
                # Check token_id bounds to ensure within vocabulary
                if token_id < 0 or token_id >= vocab_size:
                    raise ValueError(
                        f"token_id {token_id} in logit_bias contains "
                        f"out-of-vocab token id. Vocabulary size: "
                        f"{vocab_size}"
                    )
                logits[i, token_id] += bias

    torch.cuda.synchronize()
    return logits


mbs = 8
vocab_size = 32000

import random

logits = torch.rand(mbs, vocab_size, device="cuda")

bs = 5

for num_bias in [10, 50, 100]:
    logit_bias_list = []
    for _ in range(bs):
        logit_bias = {}
        for _ in range(num_bias):
            logit_bias[random.randint(0, vocab_size - 1)] = random.random()
        logit_bias_list.append(logit_bias)
    assert torch.allclose(
        apply_logits_bias_opt(logits, logit_bias_list),
        apply_logits_bias(logits, logit_bias_list),
    )
    print(f"---------num_bias: {num_bias}, batch size: {bs}---------")
    print("this PR: batched logits update")
    %timeit apply_logits_bias_opt(logits, logit_bias_list)
    print("trunk non-batched logits update")
    %timeit apply_logits_bias(logits, logit_bias_list)    
```

Output:
```
---------num_bias: 10, batch size: 5---------
this PR: batched logits update
156 µs ± 884 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
trunk non-batched logits update
368 µs ± 562 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
---------num_bias: 50, batch size: 5---------
this PR: batched logits update
229 µs ± 3.37 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
trunk non-batched logits update
1.8 ms ± 6.08 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
---------num_bias: 100, batch size: 5---------
this PR: batched logits update
309 µs ± 370 ns per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
trunk non-batched logits update
3.62 ms ± 4.67 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```