@yuguo68 2025-06-05
@houseroad , @njhill added micro benchmark, we do have significant improvement with batched update especially when the number of updates is large, e.g., 3.62 ms to 309 µs with 500 updates.